refactor(insights): break InsightSidePanelContent into SidePanel/ folder
Split the monolithic InsightSidePanelContent component into focused
section components: InsightPanelInfo, InsightPanelActions,
InsightPanelToggles, and InsightPanelDangerZone. Extract dialog helpers
into insightSidePanelDialogs. Add storybook stories for different
insight states.
