+++ frontend/src/scenes/insights/__mocks__/createInsightScene.tsx
@@ -4,6 +4,7 @@ import { router } from 'kea-router'
 import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
 import { App } from 'scenes/App'
 
+import { sceneLayoutLogic } from '~/layout/scenes/sceneLayoutLogic'
 import { useStorybookMocks } from '~/mocks/browser'
 import { InsightVizNode, Node } from '~/queries/schema/schema-general'
 import { isInsightVizNode, isLifecycleQuery, isStickinessQuery, isTrendsQuery } from '~/queries/utils'
@@ -37,11 +38,16 @@ function setLegendFilter(query: Node | null | undefined, showLegend: boolean): N
     return query
 }
 
+interface InsightStoryOptions {
+    openSidePanel?: boolean
+}
+
 let shortCounter = 0
 export function createInsightStory(
     insight: Partial<QueryBasedInsightModel>,
     mode: 'view' | 'edit' = 'view',
-    showLegend: boolean = false
+    showLegend: boolean = false,
+    options: InsightStoryOptions = {}
 ): StoryFn<typeof App> {
     const count = shortCounter++
     return function InsightStory() {
@@ -82,6 +88,9 @@ export function createInsightStory(
 
         useOnMountEffect(() => {
             router.actions.push(`/insights/${insight.short_id}${count}${mode === 'edit' ? '/edit' : ''}`)
+            if (options.openSidePanel) {
+                sceneLayoutLogic.actions.setScenePanelOpen(true)
+            }
         })
 
         return <App />
